Updated tff_patch/federated_averaging.py
