+++ tff_patch/federated_averaging.py
@@ -22,7 +22,8 @@ Communication-Efficient Learning of Deep Networks from Decentralized Data
 """
 
 import collections
-from typing import Any, Callable, Optional, Union
+from typing import Callable, Optional, Any, Union
+import warnings
 
 import tensorflow as tf
 
@@ -30,8 +31,10 @@ from tensorflow_federated.python.aggregators import factory
 from tensorflow_federated.python.common_libs import py_typecheck
 from tensorflow_federated.python.core.api import computations
 from tensorflow_federated.python.core.templates import iterative_process
+from tensorflow_federated.python.core.templates import measured_process
+from tensorflow_federated.python.learning import client_weight_lib
 from tensorflow_federated.python.learning import model as model_lib
-from tensorflow_federated.python.learning import model_utils, ClientWeighting
+from tensorflow_federated.python.learning import model_utils
 from tensorflow_federated.python.learning.framework import dataset_reduce
 from tensorflow_federated.python.tensorflow_libs import tensor_utils
 
@@ -45,15 +48,15 @@ ClientWeightFnType = Callable[[Any], tf.Tensor]
 class ClientFedAvg(optimizer_utils.ClientDeltaFn):
   """Client TensorFlow logic for Federated Averaging."""
 
-  def __init__(self,
-               model: model_lib.Model,
-               optimizer: tf.keras.optimizers.Optimizer,
-               client_weighting: Union[
-                 ClientWeighting,
-                 ClientWeightFnType] = ClientWeighting.NUM_EXAMPLES,
-               use_experimental_simulation_loop: bool = False,
-               byzantine_client_weight: int = 1_000_000,
-               attack: Optional[LocalAttack] = None):
+  def __init__(
+      self,
+      model: model_lib.Model,
+      optimizer: tf.keras.optimizers.Optimizer,
+      client_weighting: Union[client_weight_lib.ClientWeightType,
+                              ClientWeightFnType] = client_weight_lib.ClientWeighting.NUM_EXAMPLES,
+      use_experimental_simulation_loop: bool = False,
+      byzantine_client_weight: int = 1_000_000,
+      attack: Optional[LocalAttack] = None):
     """Creates the client computation for Federated Averaging.
 
     Note: All variable creation required for the client computation (e.g. model
@@ -69,20 +72,19 @@ class ClientFedAvg(optimizer_utils.ClientDeltaFn):
         provides the weight in the federated average of model deltas.
       use_experimental_simulation_loop: Controls the reduce loop function for
         input dataset. An experimental reduce loop is used for simulation.
+      byzantine_client_weight: Number of samples each Byzantine client reports.
+      attack: An optional `LocalAttack` that specifies which Byzantine attack takes place.
     """
     py_typecheck.check_type(model, model_lib.Model)
     self._model = model_utils.enhance(model)
     self._optimizer = optimizer
     py_typecheck.check_type(self._model, model_utils.EnhancedModel)
-
-    if not isinstance(client_weighting, ClientWeighting) and not callable(client_weighting):
-      raise TypeError(f'`client_weighting` must be either instance of '
-                      f'`ClientWeighting` or callable. '
+    if not isinstance(client_weighting, client_weight_lib.ClientWeighting) and not callable(client_weighting):
+      raise TypeError('`client_weighting` must be either instance of `ClientWeighting` or callable. '
                       f'Found type {type(client_weighting)}.')
     self._client_weighting = client_weighting
 
-    self._dataset_reduce_fn = dataset_reduce.build_dataset_reduce_fn(
-      use_experimental_simulation_loop)
+    self._dataset_reduce_fn = dataset_reduce.build_dataset_reduce_fn(use_experimental_simulation_loop)
     self._byzantine_client_weight = byzantine_client_weight
     self._attack = attack
 
@@ -94,9 +96,6 @@ class ClientFedAvg(optimizer_utils.ClientDeltaFn):
   def __call__(self, dataset_with_byzflag, initial_weights):
     model = self._model
     optimizer = self._optimizer
-
-    dataset, byzflag = dataset_with_byzflag
-
     tf.nest.map_structure(lambda a, b: a.assign(b), model.weights,
                           initial_weights)
 
@@ -110,62 +109,72 @@ class ClientFedAvg(optimizer_utils.ClientDeltaFn):
 
       if output.num_examples is None:
         return num_examples_sum + tf.shape(
-          output.predictions, out_type=tf.int64)[0]
+            output.predictions, out_type=tf.int64)[0]
       else:
         return num_examples_sum + tf.cast(output.num_examples, tf.int64)
 
+    dataset, byzflag = dataset_with_byzflag
+
     num_examples_sum = self._dataset_reduce_fn(
-      reduce_fn,
-      dataset,
-      initial_state_fn=lambda: tf.zeros(shape=[], dtype=tf.int64))
+        reduce_fn,
+        dataset,
+        initial_state_fn=lambda: tf.zeros(shape=[], dtype=tf.int64))
 
     weights_delta = tf.nest.map_structure(tf.subtract, model.weights.trainable,
                                           initial_weights.trainable)
     model_output = model.report_local_outputs()
 
     if byzflag:
-      # if self._attack == 'delta_to_zero':
-      #   weights_delta = tf.nest.map_structure(lambda _: -_, initial_weights.trainable)
+      # delta to zero attack
+      # weights_delta = tf.nest.map_structure(lambda _: -_, initial_weights.trainable)
       weights_delta = self._attack(weights_delta)
 
-    weights_delta, has_non_finite_delta = (tensor_utils.zero_all_if_any_non_finite(weights_delta))
+    # TODO(b/122071074): Consider moving this functionality into
+    # tff.federated_mean?
+    weights_delta, has_non_finite_delta = (
+        tensor_utils.zero_all_if_any_non_finite(weights_delta))
     # Zero out the weight if there are any non-finite values.
     if has_non_finite_delta > 0:
+      # TODO(b/176171842): Zeroing has no effect with unweighted aggregation.
       weights_delta_weight = tf.constant(0.0)
-    elif self._client_weighting is ClientWeighting.NUM_EXAMPLES:
-      weights_delta_weight = tf.cast(num_examples_sum, tf.float32)
-    elif self._client_weighting is ClientWeighting.UNIFORM:
+    elif self._client_weighting is client_weight_lib.ClientWeighting.NUM_EXAMPLES:
+      if byzflag:
+        weights_delta_weight = tf.cast(self._byzantine_client_weight, tf.float32)
+      else:
+        weights_delta_weight = tf.cast(num_examples_sum, tf.float32)
+    elif self._client_weighting is client_weight_lib.ClientWeighting.UNIFORM:
       weights_delta_weight = tf.constant(1.0)
     else:
       if byzflag:
-        # TODO generalize to things other than experiments
-        # TODO parameterize num_tokens (client weight)
+        # TODO generalize
         weights_delta_weight = self._client_weighting({'num_tokens': self._byzantine_client_weight})
       else:
         weights_delta_weight = self._client_weighting(model_output)
-
+    # TODO(b/176245976): TFF `ClientOutput` structure names are confusing.
     optimizer_output = collections.OrderedDict(num_examples=num_examples_sum)
     return optimizer_utils.ClientOutput(weights_delta, weights_delta_weight,
                                         model_output, optimizer_output)
 
 
-def default_server_optimizer_fn():
-  return tf.keras.optimizers.SGD(learning_rate=1.0)
+DEFAULT_SERVER_OPTIMIZER_FN = lambda: tf.keras.optimizers.SGD(learning_rate=1.0)
 
 
+# TODO(b/170208719): Remove `aggregation_process` after migration to
+# `model_update_aggregation_factory`.
 def build_federated_averaging_process(
     model_fn: Callable[[], model_lib.Model],
     client_optimizer_fn: Callable[[], tf.keras.optimizers.Optimizer],
     server_optimizer_fn: Callable[
-      [], tf.keras.optimizers.Optimizer] = default_server_optimizer_fn,
+        [], tf.keras.optimizers.Optimizer] = DEFAULT_SERVER_OPTIMIZER_FN,
     *,  # Require named (non-positional) parameters for the following kwargs:
-    client_weighting: Optional[Union[ClientWeighting,
-                                     ClientWeightFnType]] = None,
+    client_weighting: Optional[Union[client_weight_lib.ClientWeightType, ClientWeightFnType]] = None,
+    broadcast_process: Optional[measured_process.MeasuredProcess] = None,
+    aggregation_process: Optional[measured_process.MeasuredProcess] = None,
     model_update_aggregation_factory: Optional[
-      factory.WeightedAggregationFactory] = None,
+        factory.WeightedAggregationFactory] = None,
     use_experimental_simulation_loop: bool = False,
     byzantine_client_weight: int = 1_000_000,
-    attack: Optional[LocalAttack] = None,
+    attack: Optional[LocalAttack] = None
 ) -> iterative_process.IterativeProcess:
   """Builds an iterative process that performs federated averaging.
 
@@ -210,6 +219,12 @@ def build_federated_averaging_process(
   sophisticated federated averaging procedures may use different learning rates
   or server optimizers.
 
+  WARNING: `aggregation_process` argument is deprecated and will be removed in
+  a future version. Use `model_update_aggregation_factory` instead. See
+  https://www.tensorflow.org/federated/tutorials/tuning_recommended_aggregators
+  and https://www.tensorflow.org/federated/tutorials/custom_aggregators
+  tutorials for details of use of `tff.aggregators` module.
+
   Args:
     model_fn: A no-arg function that returns a `tff.learning.Model`. This method
       must *not* capture TensorFlow tensors or variables and use them. The model
@@ -224,6 +239,15 @@ def build_federated_averaging_process(
       `model.report_local_outputs` and returns a tensor that provides the weight
       in the federated average of model deltas. If None, defaults to weighting
       by number of examples.
+    broadcast_process: a `tff.templates.MeasuredProcess` that broadcasts the
+      model weights on the server to the clients. It must support the signature
+      `(input_values@SERVER -> output_values@CLIENT)`. If set to default None,
+      the server model is broadcast to the clients using the default
+      tff.federated_broadcast.
+    aggregation_process: a `tff.templates.MeasuredProcess` that aggregates the
+      model updates on the clients back to the server. It must support the
+      signature `({input_values}@CLIENTS-> output_values@SERVER)`. Must be
+      `None` if `model_update_aggregation_factory` is not `None.`
     model_update_aggregation_factory: An optional
       `tff.aggregators.WeightedAggregationFactory` or
       `tff.aggregators.UnweightedAggregationFactory` that constructs
@@ -234,7 +258,7 @@ def build_federated_averaging_process(
         input dataset. An experimental reduce loop is used for simulation.
         It is currently necessary to set this flag to True for performant GPU
         simulations.
-    byzantine_client_weight: Number of samples each Byzantine client reports
+    byzantine_client_weight: Number of samples each Byzantine client reports.
     attack: An optional `LocalAttack` that specifies which Byzantine attack takes place
 
   Returns:
@@ -244,22 +268,36 @@ def build_federated_averaging_process(
   if isinstance(model_update_aggregation_factory,
                 factory.UnweightedAggregationFactory):
     if client_weighting is None:
-      client_weighting = ClientWeighting.UNIFORM
-    elif client_weighting is not ClientWeighting.UNIFORM:
+      client_weighting = client_weight_lib.ClientWeighting.UNIFORM
+    elif client_weighting is not client_weight_lib.ClientWeighting.UNIFORM:
       raise ValueError('Cannot use non-uniform client weighting with '
                        'unweighted aggregation.')
-  elif not client_weighting:
-    client_weighting = ClientWeighting.NUM_EXAMPLES
+  elif client_weighting is None:
+    client_weighting = client_weight_lib.ClientWeighting.NUM_EXAMPLES
+
+  if aggregation_process is not None:
+    warnings.warn(
+        'The aggregation_process argument to '
+        'tff.learning.build_federated_averaging_process is deprecated and will '
+        'be removed in a future version. Use model_update_aggregation_factory '
+        'instead. See '
+        'https://www.tensorflow.org/federated/tutorials/tuning_recommended_aggregators'
+        ' and '
+        'https://www.tensorflow.org/federated/tutorials/custom_aggregators '
+        'tutorials for details of use of tff.aggregators module.',
+        DeprecationWarning)
 
   def client_fed_avg(model_fn: Callable[[], model_lib.Model]) -> ClientFedAvg:
     return ClientFedAvg(model_fn(), client_optimizer_fn(), client_weighting,
                         use_experimental_simulation_loop, byzantine_client_weight, attack)
 
   iter_proc = optimizer_utils.build_model_delta_optimizer_process(
-    model_fn,
-    model_to_client_delta_fn=client_fed_avg,
-    server_optimizer_fn=server_optimizer_fn,
-    model_update_aggregation_factory=model_update_aggregation_factory)
+      model_fn,
+      model_to_client_delta_fn=client_fed_avg,
+      server_optimizer_fn=server_optimizer_fn,
+      broadcast_process=broadcast_process,
+      aggregation_process=aggregation_process,
+      model_update_aggregation_factory=model_update_aggregation_factory)
 
   server_state_type = iter_proc.state_type.member
